Bug fix for Chinese orginal language
Bug where a Chinese original language movie would be set as und or undefined because of the return value from TMBD as 'cn'.  The iso-language expects 'zh' for Chinese.
Also the array is unnecessary given TMDB should only every report a single language and is not reliant on imdb multiple languages.