Fixed an out of memory issue that prevented .ts files from being processed by removing the framerate portion of the video filter command when processing a .ts file.
Removed a bunch of variables used for building the ffmpeg command and
moved the actual strings down to where the logic is used. Reducing
plugin file size and variable ussage.

Fixed a minor bug related to changing the video height where it would
always use the 1080 value rather than using the value used in the input.