Fix error with ts files and cleanup (#271)
* Fixed the NaN issue with BitRates. Some files had bitrate missing in the
MediaInfo data so pulling from the "extra" section for those. Also
resolved an issue with the way the findMediaInfoItem function wasn't
working with subtitle tracks (well it wasn't working for all files due
to the general section because of a previous attempted subtitle fix but
this should be much much much better and work in most all cases).

* Updating to comply with eslint. Unabe to get stream matching function to
work so changed eqeqeq to warning.

* Re-add string

* parseInt

* Merge remote-tracking branch 'refs/remotes/origin/master'

* Modified variable names to use the lowerCamelCase naming convetion

* Fixed eslint issue

* Fixed an out of memory issue that prevented .ts files from being
processed by removing the framerate portion of the video filter
command when processing a .ts file.

Removed a bunch of variables used for building the ffmpeg command and
moved the actual strings down to where the logic is used. Reducing
plugin file size and variable ussage.

Fixed a minor bug related to changing the video height where it would
always use the 1080 value rather than using the value used in the input.

* Fixed issue with template line.

* Removed the commented code. Cleaned up other comments to improve
readability.

Co-authored-by: HaveAGitGat <43864057+HaveAGitGat@users.noreply.github.com>