Reworked to ensure we get accurate video bitrate
Use mkvpropedit to get new stats on video files.
Ignore files if they are less than 7 days old as they "should" be upto date
If a file is converted to HEVC and we have HEVC reprocess enabled then we encode in very old dates to ensure on next pass we get new stats again.
Also a bunch of cleanup of strings to ensure they appear in the log as expected