Updated transfer_characteristics to color_transfer
Current implementation is looking for "transfer_characteristics" within ffprobe data as an attribute that has to match to determine if a file is HDR or not. However, current ffprobe is not showing this attribute on HDR files. I am proposing changing this attribute to "color_transfer" which does show up on HDR files.