Update Tdarr_Plugin_bsh1_Boosh_FFMPEG_QSV_HEVC.js
Fixed up an issue with HW decoding. If the -hwaccel qsv is set then ffmpeg seems to automatically enable hw decoding even when we don't want it. There's been several reports of issues with certain files & I think this is the cause. I've tested and that appears to resolve the transcode error