Add files via upload
These 2 files will back up and restore the metadata dates, but unlike the original ones these give the option of where to place the files and then creates a subfolder from the filename and puts the .dates file inside of it.  Can be used in conjunction with Flow Plugin checkFileExistsModified.