Create Tdarr_Plugin_jordy_Remove_Audio_By_Codec_Channels.js
This plugin will remove audio tracks from a file based on the codec, channel configuration, and language. It will only remove audio tracks that match the specified codec and channel configuration

For example, if you specify 'aac' and '5.1', it will remove all audio tracks that are AAC and have 6 channels.